+++ ChangeLog.md
@@ -1,7 +1,13 @@
 # Changelog
 
 ## Version [0.9.0](https://github.com/cedx/sql.net/compare/v0.8.0...v0.9.0)
+- Removed the `CommandExtensions` and `ListExtensions` classes.
+- Renamed the `DataMapper` class to `Mapper`.
+- Renamed the `QueryOptions` record to `CommandOptions`.
+- Renamed the `New-DataMapper` cmdlet to `New-Mapper`.
+- Merged the `-Parameters` and `-PositionalParameters` cmdlet parameters.
 - Added the `New-Transaction`, `Approve-Transaction` and `Deny-Transaction` cmdlets.
+- Added the `-Precision`, `-Scale` and `-Size` parameters to the `New-Parameter` cmdlet.
 - Added the `-Transaction` parameter to most cmdlets.
 
 ## Version [0.8.0](https://github.com/cedx/sql.net/compare/v0.7.0...v0.8.0)