+++ ChangeLog.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## Version [0.11.0](https://github.com/cedx/sql.net/compare/v0.10.0...v0.11.0)
+
 ## Version [0.10.0](https://github.com/cedx/sql.net/compare/v0.9.0...v0.10.0)
 - Added the `TableInfo` and `ColumnInfo` classes.
 - The `Get-First` and `Get-Single` cmdlets now use a non-terminating error.